+++ src/promptflow-core/promptflow/executor/flow_executor.py
@@ -205,14 +205,18 @@ class FlowExecutor:
         if hasattr(flow_file, "__call__") or inspect.isfunction(flow_file):
             from ._script_executor import ScriptExecutor
 
-            return ScriptExecutor(flow_file, storage=storage)
+            return ScriptExecutor(flow_file, connections=connections, storage=storage)
         if not isinstance(flow_file, (Path, str)):
             raise NotImplementedError("Only support Path or str for flow_file.")
         if is_flex_flow(flow_path=flow_file, working_dir=working_dir):
             from ._script_executor import ScriptExecutor
 
             return ScriptExecutor(
-                flow_file=Path(flow_file), working_dir=working_dir, storage=storage, init_kwargs=init_kwargs
+                flow_file=Path(flow_file),
+                connections=connections,
+                working_dir=working_dir,
+                storage=storage,
+                init_kwargs=init_kwargs,
             )
         elif is_prompty_flow(file_path=flow_file):
             from ._prompty_executor import PromptyExecutor
@@ -863,7 +867,7 @@ class FlowExecutor:
         context: FlowExecutionContext,
         stream=False,
     ):
-        with self._start_flow_span(inputs) as span, self._record_keyboard_interrupt_to_span(span):
+        with self._start_flow_span(inputs) as span, self._record_cancellation_exceptions_to_span(span):
             output, nodes_outputs = await self._traverse_nodes_async(inputs, context)
             #  TODO: Also stringify async generator output
             output = self._stringify_generator_output(output) if not stream else output
@@ -878,17 +882,17 @@ class FlowExecutor:
         context: FlowExecutionContext,
         stream=False,
     ):
-        with self._start_flow_span(inputs) as span, self._record_keyboard_interrupt_to_span(span):
+        with self._start_flow_span(inputs) as span, self._record_cancellation_exceptions_to_span(span):
             output, nodes_outputs = self._traverse_nodes(inputs, context)
             output = self._stringify_generator_output(output) if not stream else output
             self._exec_post_process(inputs, output, nodes_outputs, run_info, run_tracker, span, stream)
             return output, extract_aggregation_inputs(self._flow, nodes_outputs)
 
     @contextlib.contextmanager
-    def _record_keyboard_interrupt_to_span(self, span: Span):
+    def _record_cancellation_exceptions_to_span(self, span: Span):
         try:
             yield
-        except KeyboardInterrupt as ex:
+        except (KeyboardInterrupt, asyncio.CancelledError) as ex:
             if span.is_recording():
                 span.record_exception(ex)
                 span.set_status(StatusCode.ERROR, "Execution cancelled.")