+++ src/promptflow-devkit/CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Improvements
 - Interactive browser credential is excluded by default when using Azure AI connections, user could set `PF_NO_INTERACTIVE_LOGIN=False` to enable it.
 
+### Bugs Fixed
+- Fix the issue that import error will be raised after downgrading promptflow from >=1.10.0 to <1.8.0.
+- Fix the issue that `pf flow serve` is broken with exception `NotADirectoryError`.
+
 ## v1.10.0 (2024.04.26)
 
 ### Features Added